Kind of works
There is one more thing - the issue described here: https://www.reddit.com/r/adventofcode/comments/a6dp5v/2018_day_15_part1_cant_get_the_right_answer/ebtzthk/
Basically it should first select "in range" cell and then shortest path to it. Not the shortest path to the enemy.
It should be enough to create distmap from the unit that moves - find the best "in range" and then use the current impl - just calculate distmap for that "in rance" cell instead of for the target.